etl/requirements.txt
dnspython==2.2.1
numpy==1.24.0
pandas==1.5.2
pymongo==4.3.3
python-dateutil==2.8.2
python-dotenv==1.0.0
pytz==2022.7
six==1.16.0
